Refactor routing and component imports in App.js; update Box1 export syntax; modify Footer layout and logo text.
