+++ src/layout/Footer.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import  guilty from '../../Assets/guilty.jpg';
+
 function Footer(){
     return(
 <footer className="footer"> 
-    <div  className="log">
-    <img src={guilty.jpg} alt="Guilty logo"/>
+    <div className="log">
+        <div className="logo-text">Guilty or Not</div>
     </div>
 </footer>
     );
